+++ src/app/page.tsx
@@ -2,17 +2,18 @@
 import { Status, Wrapper } from '@googlemaps/react-wrapper';
 import CarFfeineMap from '@/components/google-maps/map/CarFfeineMap';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import Loading from '@ui/Loading';
 
 const queryClient = new QueryClient();
 
 const render = (status: Status) => {
   switch (status) {
+    case Status.LOADING:
+      return <Loading />;
     case Status.FAILURE:
       return <>에러 발생</>;
-    case Status.LOADING:
-      return <div>로딩중</div>;
     case Status.SUCCESS:
-      return <CarFfeineMap/>;
+      return <CarFfeineMap />;
   }
 };
