+++ src/components/login-page/GoogleLogin.tsx
@@ -1,4 +1,4 @@
-import Loading from 'components/ui/Loading';
+import Loading from '@components/ui/Loading';
 
 import { useEffect, useState } from 'react';
