+++ src/types/index.ts
@@ -2,4 +2,4 @@ export * from './stations';
 export * from './chargers';
 export * from './congestion';
 export * from './style';
-export { DisplayPosition } from '@type/map';
+export type { DisplayPosition } from '@type/map';
